API_Youtube/requirements.txt
asgiref==3.3.1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.4
django-crontab==0.7.1
django-filter==2.4.0
djangorestframework==3.12.2
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
protobuf==3.15.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
requests==2.25.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
gunicorn==20.0.4
apscheduler
